cpu/se3208/se3208dis.cpp, */vrender0.cpp: Clean up a few details: (#14715)
* cpu/se3208/se3208dis.cpp: Avoid reserved name for type.
* sound/vrender0.cpp: Simplify casting.
* video/vrender0.cpp: Use pal5bit/pal6bit functions for color field extraction.

+++ src/devices/sound/vrender0.cpp
@@ -532,7 +532,7 @@ void vr0sound_device::render_audio(sound_stream &stream)
 				if (channel.modes & MODE_8BIT)   //8bit
 				{
 					sample = channel.cache->read_byte(channel.cur_saddr >> 9);
-					sample = s16(s8(sample & 0xff) << 8);
+					sample = s16((sample & 0xff) << 8);
 				}
 				else                //16bit
 				{
